backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (PostgreSQL async)
sqlalchemy==2.0.25
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# External APIs
ccxt==4.2.22
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
cryptography==42.0.7

# Testing
pytest==8.2.2
pytest-asyncio==0.23.8

# Migrations
alembic==1.13.2
